refactor: drawable한 요소들을 path와 shape로 나눔
