+++ app/src/main/java/woowacourse/paint/main/PaintBoard.kt
@@ -12,15 +12,16 @@ import woowacourse.paint.model.BrushSize
 import woowacourse.paint.model.DrawMode
 import woowacourse.paint.model.DrawableHistory
 import woowacourse.paint.model.PaintColor
-import woowacourse.paint.model.drawable.DrawableCircle
 import woowacourse.paint.model.drawable.DrawableElement
-import woowacourse.paint.model.drawable.DrawableEraser
-import woowacourse.paint.model.drawable.DrawablePath
-import woowacourse.paint.model.drawable.DrawableSquare
+import woowacourse.paint.model.drawable.path.DrawableEraser
+import woowacourse.paint.model.drawable.path.DrawableLine
+import woowacourse.paint.model.drawable.path.DrawablePath
+import woowacourse.paint.model.drawable.shape.DrawableCircle
+import woowacourse.paint.model.drawable.shape.DrawableSquare
 
 class PaintBoard constructor(context: Context, attrs: AttributeSet) : View(context, attrs) {
     private val pathHistory = DrawableHistory()
-    private var currentDraw: DrawableElement = DrawablePath(paint = Paint())
+    private var currentDraw: DrawableElement = DrawableLine(paint = Paint())
 
     init {
         setLayerType(LAYER_TYPE_HARDWARE, null)
@@ -63,7 +64,7 @@ class PaintBoard constructor(context: Context, attrs: AttributeSet) : View(conte
             strokeWidth = currentDraw.paint.strokeWidth
         }
         currentDraw = when (mode) {
-            DrawMode.BRUSH -> DrawablePath(paint = newPaint)
+            DrawMode.BRUSH -> DrawableLine(paint = newPaint)
             DrawMode.SQUARE -> DrawableSquare(paint = newPaint)
             DrawMode.CIRCLE -> DrawableCircle(paint = newPaint)
             DrawMode.ERASER -> DrawableEraser(paint = newPaint)
@@ -73,8 +74,6 @@ class PaintBoard constructor(context: Context, attrs: AttributeSet) : View(conte
     fun setBrushSize(size: BrushSize) {
         if (currentDraw is DrawablePath) {
             currentDraw = (currentDraw as DrawablePath).changeBrushSize(size)
-        } else if (currentDraw is DrawableEraser) {
-            currentDraw = (currentDraw as DrawableEraser).changeBrushSize(size)
         }
     }
