+++ app/src/main/java/woowacourse/paint/model/drawable/shape/DrawableShape.kt
@@ -1,0 +1,5 @@
+package woowacourse.paint.model.drawable.shape
+
+import woowacourse.paint.model.drawable.DrawableElement
+
+interface DrawableShape : DrawableElement
